More work on simplifying the gate count sample, updating to beta package.
